requirements.txt
fastapi==0.112.0
uvicorn==0.30.3
scikit-learn==1.5.1
numpy>=1.24
pandas>=2.0.0
joblib>=1.3.0
pydantic>=2.6.0
requests>=2.31.0
pyyaml>=6.0
